feat: return on post request response from new template
