+++ src/db/db_models.rs
@@ -1,12 +1,6 @@
-// Generated by diesel_ext
-
-#![allow(unused)]
-#![allow(clippy::all)]
-
-use crate::db::schema::notification_templates::dsl::notification_templates;
 use chrono::NaiveDateTime;
 use diesel::{Insertable, Queryable, Selectable};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 
 #[derive(Queryable, Debug)]
 #[diesel(table_name = notification_recipients)]
@@ -19,7 +13,7 @@ pub struct NotificationRecipient {
     pub updated_at: Option<NaiveDateTime>,
 }
 
-#[derive(Selectable, Queryable, Debug)]
+#[derive(Selectable, Queryable, Debug, Serialize)]
 #[diesel(table_name = super::schema::notification_templates)]
 pub struct NotificationTemplate {
     pub id: i32,
