+++ src/errors.rs
@@ -1,28 +1,7 @@
 use axum::extract::rejection::JsonRejection;
 use axum::response::{IntoResponse, Response};
 use serde::Serialize;
-use crate::handlers::post_templates::JsonExtractor;
-
-// pub struct AppError(anyhow::Error);
-//
-// impl IntoResponse for AppError {
-//     fn into_response(self) -> Response {
-//         (
-//             StatusCode::INTERNAL_SERVER_ERROR,
-//             format!("Something went wrong: {}", self.0),
-//         )
-//             .into_response()
-//     }
-// }
-//
-// impl<E> From<E> for AppError
-// where
-//     E: Into<anyhow::Error>,
-// {
-//     fn from(err: E) -> Self {
-//         Self(err.into())
-//     }
-// }
+pub use crate::handlers::JsonExtractor;
 
 pub enum AppError {
     JsonRejection(JsonRejection),
